+++ src/client.spec.ts
@@ -199,7 +199,7 @@ describe('client', () => {
         const result = await client.verifyPasswordlessLogin(VERIFY_URL);
         expect(result.verificationStatus).toBe(TokenVerificationStatus.USED);
       });
-      
+
       it('should throw when bad request status received', async () => {
         fetchMock.mockResponse(
           JSON.stringify({
@@ -213,6 +213,23 @@ describe('client', () => {
           client.verifyPasswordlessLogin(VERIFY_URL)
         ).rejects.toThrow();
       });
+    });
+  });
+
+  describe('github', () => {
+    beforeEach(() => {
+      client.setOptions({ baseUrl: BASE_URL });
+    });
+
+    it('should return github login url', () => {
+      expect(client.getGithubLoginUrl()).toBe(`${BASE_URL}/auth/github`);
+    });
+
+    it('should return github login url with urlencoded returnTo query parameter', () => {
+      const returnTo = 'https://superface.dev/login/callback';
+      expect(client.getGithubLoginUrl(returnTo)).toBe(
+        `${BASE_URL}/auth/github?return_to=${encodeURIComponent(returnTo)}`
+      );
     });
   });
 });